- added an ExternalStoreRelation which extends InMemoryAppendableRelation. Corresponding changes to SnappyContext. - Passing the conn string and the driver to the constructor of ExternalStoreRelation.
Remaining work is to use the above to obtain connection and push/pull data to/from store.
Adding a method to take store as an argument such that the InMemoryRelation created for the sample table / or exact table is of type ExternalStoreRelation than InMemoryAppendable or InMemoryRelation
as the case may be.